Move mocks folder to domain entity folder
