+++ src/battle/hooks/__tests__/editBattle.test.tsx
@@ -4,7 +4,7 @@ import {
   llucmajorBattle,
   montjuicBattle,
   editedAlmansaBattle,
-} from "../../../fixtures";
+} from "../../fixtures";
 import setupStore from "../../../store/setupStore";
 import type { BattlesInfoState } from "../../slice/types";
 import { renderHook } from "@testing-library/react";
