+++ src/battle/hooks/__tests__/toggleBattleWinner.test.tsx
@@ -3,11 +3,7 @@ import { act, renderHook } from "@testing-library/react";
 import useBattles from "../useBattles";
 import setupStore from "../../../store/setupStore";
 import type { BattlesInfoState } from "../../slice/types";
-import {
-  almansaBattle,
-  llucmajorBattle,
-  montjuicBattle,
-} from "../../../fixtures";
+import { almansaBattle, llucmajorBattle, montjuicBattle } from "../../fixtures";
 
 describe("Given the toggleBattleWinner method from useBattle hook", () => {
   describe("When it receives the ID for Battle of Almansa", () => {
