+++ src/battle/mocks/node.ts
