+++ src/setupTest.ts
@@ -1,6 +1,6 @@
-import "@testing-library/jest-dom/vitest";
 import { beforeAll, afterEach, afterAll } from "vitest";
-import { server } from "./mocks/node";
+import { server } from "./battle/mocks/node";
+import "@testing-library/jest-dom/vitest";
 
 beforeAll(() => server.listen());
 afterEach(() => server.resetHandlers());
